4/1/2018 - Project3
Created the Project3 folder, I also believe I fixed the ace-builds empty folder issue. The whole project is finished. April Fools no it's not. You must feel like such a fool. Now that we got that out of the way, I've been meaning to start actually working on this for a while but things keep just popping up, so this should be fun. Now, to bed!